+++ scripts/post-release.sh
@@ -5,17 +5,11 @@
 
 set -eux
 
-SCRIPT_DIR="$( cd "$( dirname "${BASH_SOURCE[0]}" )" && pwd )"
-cd $SCRIPT_DIR/..
-
-OLD_VERSION="${1}"
-NEW_VERSION="${2}"
-
 git checkout main
 GRADLE_FILEPATH="gradle.properties"
 
 # Add a new unreleased entry in the changelog
-sed -i "" 's/# Changelog/# Changelog\n\n## Unreleased/' CHANGELOG.md
+perl -pi -e 's/# Changelog/# Changelog\n\n## Unreleased/' CHANGELOG.md
 
 # Increment `versionName` and make it a snapshot
 # Incrementing the version name before the release (`bump-version.sh`) sets a
@@ -32,7 +26,7 @@ version_digit_to_bump="$( awk "/$VERSION_NAME_PATTERN/" $GRADLE_FILEPATH | egrep
 # Using `*` instead of `+` for compatibility. The result is the same,
 # since the version to be bumped is extracted using `+`.
 new_version="$( echo $version | sed "s/[0-9]*$/$version_digit_to_bump/g" )"
-sed -i "" -e "s/$VERSION_NAME_PATTERN=.*$/$VERSION_NAME_PATTERN=$new_version-SNAPSHOT/g" $GRADLE_FILEPATH
+perl -pi -e "s/$VERSION_NAME_PATTERN=.*$/$VERSION_NAME_PATTERN=$new_version-SNAPSHOT/g" $GRADLE_FILEPATH
 
 git add .
 git commit -m "Prepare $new_version"