+++ sentry-android-core/src/main/java/io/sentry/android/core/ActivityFramesTracker.java
@@ -12,7 +12,12 @@ import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
 import org.jetbrains.annotations.TestOnly;
 
-final class ActivityFramesTracker {
+/**
+ * A class that tracks slow and frozen frames using the FrameMetricsAggregator class from
+ * androidx.core package. It also checks if the FrameMetricsAggregator class is available at
+ * runtime.
+ */
+public final class ActivityFramesTracker {
 
   private @Nullable FrameMetricsAggregator frameMetricsAggregator = null;
   private boolean androidXAvailable = true;
@@ -20,7 +25,7 @@ final class ActivityFramesTracker {
   private final @NotNull Map<SentryId, Map<String, @NotNull MeasurementValue>>
       activityMeasurements = new ConcurrentHashMap<>();
 
-  ActivityFramesTracker(final @NotNull LoadClass loadClass) {
+  public ActivityFramesTracker(final @NotNull LoadClass loadClass) {
     androidXAvailable = checkAndroidXAvailability(loadClass);
     if (androidXAvailable) {
       frameMetricsAggregator = new FrameMetricsAggregator();
@@ -47,7 +52,7 @@ final class ActivityFramesTracker {
   }
 
   @SuppressWarnings("NullAway")
-  synchronized void addActivity(final @NotNull Activity activity) {
+  public synchronized void addActivity(final @NotNull Activity activity) {
     if (!isFrameMetricsAggregatorAvailable()) {
       return;
     }
@@ -55,7 +60,8 @@ final class ActivityFramesTracker {
   }
 
   @SuppressWarnings("NullAway")
-  synchronized void setMetrics(final @NotNull Activity activity, final @NotNull SentryId sentryId) {
+  public synchronized void setMetrics(
+      final @NotNull Activity activity, final @NotNull SentryId sentryId) {
     if (!isFrameMetricsAggregatorAvailable()) {
       return;
     }
@@ -112,7 +118,7 @@ final class ActivityFramesTracker {
   }
 
   @Nullable
-  synchronized Map<String, @NotNull MeasurementValue> takeMetrics(
+  public synchronized Map<String, @NotNull MeasurementValue> takeMetrics(
       final @NotNull SentryId sentryId) {
     if (!isFrameMetricsAggregatorAvailable()) {
       return null;
@@ -125,7 +131,7 @@ final class ActivityFramesTracker {
   }
 
   @SuppressWarnings("NullAway")
-  synchronized void stop() {
+  public synchronized void stop() {
     if (isFrameMetricsAggregatorAvailable()) {
       frameMetricsAggregator.stop();
     }