+++ sentry-android-core/src/main/java/io/sentry/android/core/AndroidOptionsInitializer.java
@@ -28,6 +28,7 @@ import org.jetbrains.annotations.Nullable;
  * Android Options initializer, it reads configurations from AndroidManifest and sets to the
  * SentryOptions. It also adds default values for some fields.
  */
+@SuppressWarnings("Convert2MethodRef") // older AGP versions do not support method references
 final class AndroidOptionsInitializer {
 
   /** private ctor */
@@ -156,12 +157,13 @@ final class AndroidOptionsInitializer {
       options.addIntegration(
           new ActivityLifecycleIntegration(
               (Application) context, buildInfoProvider, activityFramesTracker));
+      options.addIntegration(new UserInteractionIntegration((Application) context, loadClass));
     } else {
       options
           .getLogger()
           .log(
               SentryLevel.WARNING,
-              "ActivityBreadcrumbsIntegration needs an Application class to be installed.");
+              "ActivityLifecycle and UserInteraction Integrations need an Application class to be installed.");
     }
     options.addIntegration(new AppComponentsBreadcrumbsIntegration(context));
     options.addIntegration(new SystemEventsBreadcrumbsIntegration(context));