+++ sentry-android-core/src/main/java/io/sentry/android/core/SentryAndroidOptions.java
@@ -17,7 +17,7 @@ public final class SentryAndroidOptions extends SentryOptions {
    */
   private boolean anrEnabled = true;
 
-  /** ANR Timeout internal in Millis Default is 5000 = 5s Used by AnrIntegration */
+  /** ANR Timeout interval in Millis Default is 5000 = 5s Used by AnrIntegration */
   private long anrTimeoutIntervalMillis = 5000;
 
   /** Enable or disable ANR on Debug mode Default is disabled Used by AnrIntegration */
@@ -37,6 +37,9 @@ public final class SentryAndroidOptions extends SentryOptions {
 
   /** Enable or disable automatic breadcrumbs for App Components Using ComponentCallbacks */
   private boolean enableAppComponentBreadcrumbs = true;
+
+  /** Enable or disable automatic breadcrumbs for User interactions Using Window.Callback */
+  private boolean enableUserInteractionBreadcrumbs = true;
 
   /**
    * Enables the Auto instrumentation for Activity lifecycle tracing.
@@ -196,6 +199,14 @@ public final class SentryAndroidOptions extends SentryOptions {
     this.enableAppComponentBreadcrumbs = enableAppComponentBreadcrumbs;
   }
 
+  public boolean isEnableUserInteractionBreadcrumbs() {
+    return enableUserInteractionBreadcrumbs;
+  }
+
+  public void setEnableUserInteractionBreadcrumbs(boolean enableUserInteractionBreadcrumbs) {
+    this.enableUserInteractionBreadcrumbs = enableUserInteractionBreadcrumbs;
+  }
+
   /**
    * Enable or disable all the automatic breadcrumbs
    *
@@ -206,6 +217,7 @@ public final class SentryAndroidOptions extends SentryOptions {
     enableAppComponentBreadcrumbs = enable;
     enableSystemEventBreadcrumbs = enable;
     enableAppLifecycleBreadcrumbs = enable;
+    enableUserInteractionBreadcrumbs = enable;
   }
 
   /**