+++ sentry-samples/sentry-samples-android/src/main/java/io/sentry/samples/android/MainActivity.java
@@ -21,8 +21,11 @@ import java.io.Writer;
 import java.util.Calendar;
 import java.util.Collections;
 import java.util.Locale;
+import timber.log.Timber;
 
 public class MainActivity extends AppCompatActivity {
+
+  private int crashCount = 0;
 
   @Override
   protected void onCreate(Bundle savedInstanceState) {
@@ -160,8 +163,19 @@ public class MainActivity extends AppCompatActivity {
         view -> SampleFragment.newInstance().show(getSupportFragmentManager(), null));
 
     binding.openThirdFragment.setOnClickListener(
+        view -> startActivity(new Intent(this, ThirdActivityFragment.class)));
+
+    binding.openGesturesActivity.setOnClickListener(
+        view -> startActivity(new Intent(this, GesturesActivity.class)));
+
+    binding.testTimberIntegration.setOnClickListener(
         view -> {
-          startActivity(new Intent(this, ThirdActivityFragment.class));
+          crashCount++;
+          Timber.i("Some info here");
+          Timber.e(
+              new RuntimeException("Uncaught Exception from Java."),
+              "Something wrong happened %d times",
+              crashCount);
         });
 
     setContentView(binding.getRoot());